feat: test.sh removing .git folder

+++ test.sh
@@ -38,50 +38,54 @@ echo -e "Destination folder  :\t$destFolder\n"
 
 # ------------- PARENTs EXISTENCE TEST -------------
 
-find $DEV_ENV -mindepth 1 -maxdepth 1 -type d -path '.git' | while read -r parentFolder; do
-parentName=$(basename $parentFolder)
-destParentPath="$destFolder/$parentName"
+find $DEV_ENV -mindepth 1 -maxdepth 1 -name '.git' -prune -o -print | while read -r parentFolder; do
+    if [[ -d $parentFolder ]]; then
+        parentName=$(basename $parentFolder)
+        destParentPath="$destFolder/$parentName"
 
-if [[ ! -d $destParentPath ]]; then
-    echo "Creating: \"$parentName\" at \"$destFolder\""
-    mkdir $destParentPath
-fi
+        if [[ ! -d $destParentPath ]]; then
+            echo "Creating: \"$parentName\" at \"$destFolder\""
+            mkdir $destParentPath
+        fi
+    fi
 done
 
 # -------------- CHILD FOLDERS PAST PARENT EXISTENCE TEST ---------
 
 # For example: folder="$DEV_ENV/.../nvim" with $DEV_ENV being most of the time: `pwd`
 # So: folder="/home/$USER/dev/env/.config/nvim
-find $DEV_ENV -mindepth 2 -type d -path '.git'| while read -r folder; do
-    # aka: nvim
+find $DEV_ENV -mindepth 2 -type d | while read -r folder; do
+    # e.g.: nvim
     folder_name=$(basename $folder)
 
     # Path without $DEV_ENV
-    # aka: .config/nvim
+    # e.g.: .config/nvim
     relative_path=$(echo $folder | sed "s|^$DEV_ENV/||")
+    if [[ "$relative_path" != *".git"* ]]; then
 
-    # Name of the root folder holding the folder "folder"
-    # aka: .config
-    rootFolder=$(echo $relative_path | sed "s|/$folder_name||")
+        # Name of the root folder holding the folder "folder"
+        # e.g.: .config
+        rootFolder=$(echo $relative_path | sed "s|/$folder_name||")
 
-    # aka: /home/$USER/.config/nvim
-    homePath="$destFolder/$relative_path"
+        # e.g.: /home/$USER/.config/nvim
+        homePath="$destFolder/$relative_path"
 
-    # aka: /home/$USER/.config
-    homeRootPath="$destFolder/$rootFolder"
-    # echo -e "${RED}Removing: rm -rf $destFolder/$relative_path${WHITE}"
-    if [[ -d $homePath ]]; then
-        rm -rf $homePath
+        # e.g.: /home/$USER/.config
+        homeRootPath="$destFolder/$rootFolder"
+        # echo -e "${RED}Removing: rm -rf $destFolder/$relative_path${WHITE}"
+        if [[ -d $homePath ]]; then
+            rm -rf $homePath
+        fi
+
+        # echo -e "${GREEN}Copying env: $folder_name ---> $homeRootPath${WHITE}"
+        mkdir -p "$homeRootPath" &>/dev/null
+        cp -rf "$folder" "$homeRootPath"
     fi
-
-    # echo -e "${GREEN}Copying env: $folder_name ---> $homeRootPath${WHITE}"
-    mkdir -p "$homeRootPath" >/dev/null 2>/dev/null
-    cp -rf "$folder" "$homeRootPath"
 done
 
 # ----------- FILES COVERAGE -----------
 
-find $DEV_ENV -maxdepth 1 -type f -path '.git' | while read -r file; do
+find $DEV_ENV -maxdepth 1 -type f | while read -r file; do
     filename=$(basename $file)
     destPath="$destFolder/$filename"
     if [[ -f $destPath ]]; then
